requirements.txt
flask
discord
python-dotenv
waitress
sans